A script to calculate the 'score' of your algo
The score, a number between 0.0 (bad) and 1.0 (excellent), is the F1-score of your classification algorithm which turns contact maps into binary images.  In such a binary image, 1 indicates a loop and 0 a 'non-loop'. Becacuse there are much fewer 1s than 0s, the binary image is encoded by the x- y- position of the 1s. This type of enconding is also commonly used by sparse matrix algorithms.